# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Smart Commit Backend",
    "description": "Smart Commit Helper Backend — multi-tenant team & commit management API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户注册",
        "description": "Register a new user account.\n\n- Email must be unique and valid.\n- Password must be at least 8 characters with uppercase, lowercase, and digit.\n- Returns the new user's id, email, and name (never the password hash).",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "用户登录",
        "description": "Authenticate a user and return JWT tokens.\n\n- Accepts email and password.\n- Returns access_token (15 min) and refresh_token (7 days).\n- On failure, returns 401 with a generic message (no email enumeration).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "刷新 Access Token",
        "description": "Exchange a valid refresh token for a new access token.\n\n- Validates the refresh token signature and expiry.\n- Enforces one-time use via a blacklist (rotation).\n- Replay of an old token invalidates the entire token family.\n- Expired refresh token returns 401 with \"请重新登录\".",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/commits": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "提交列表",
        "description": "获取提交记录列表，支持分页和筛选。",
        "operationId": "list_commits_api_v1_commits_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按成员筛选",
              "title": "Member Id"
            },
            "description": "按成员筛选"
          },
          {
            "name": "repository",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按仓库筛选",
              "title": "Repository"
            },
            "description": "按仓库筛选"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "起始日期",
              "title": "Start Date"
            },
            "description": "起始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "结束日期",
              "title": "End Date"
            },
            "description": "结束日期"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_CommitResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commits/{commit_id}": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "提交详情",
        "description": "获取单条提交记录详情。",
        "operationId": "get_commit_api_v1_commits__commit_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "commit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Commit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commits/workload/{member_id}": {
      "get": {
        "tags": [
          "commits"
        ],
        "summary": "成员工作量统计",
        "description": "获取指定成员的工作量统计。",
        "operationId": "get_workload_api_v1_commits_workload__member_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "起始日期",
              "title": "Start Date"
            },
            "description": "起始日期"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "结束日期",
              "title": "End Date"
            },
            "description": "结束日期"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommitDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "sha": {
            "type": "string",
            "title": "Sha"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "lines_added": {
            "type": "integer",
            "title": "Lines Added"
          },
          "lines_deleted": {
            "type": "integer",
            "title": "Lines Deleted"
          },
          "ai_percentage": {
            "type": "integer",
            "title": "Ai Percentage"
          },
          "committed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Committed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "member_id",
          "sha",
          "message",
          "repository",
          "branch",
          "lines_added",
          "lines_deleted",
          "ai_percentage",
          "committed_at"
        ],
        "title": "CommitDetailResponse",
        "description": "Response schema for commit detail (includes team_id)."
      },
      "CommitResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "sha": {
            "type": "string",
            "title": "Sha"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "repository": {
            "type": "string",
            "title": "Repository"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "lines_added": {
            "type": "integer",
            "title": "Lines Added"
          },
          "lines_deleted": {
            "type": "integer",
            "title": "Lines Deleted"
          },
          "ai_percentage": {
            "type": "integer",
            "title": "Ai Percentage"
          },
          "committed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Committed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "sha",
          "message",
          "repository",
          "branch",
          "lines_added",
          "lines_deleted",
          "ai_percentage",
          "committed_at"
        ],
        "title": "CommitResponse",
        "description": "Response schema for a single commit."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "description": "User email address",
            "examples": [
              "user@example.com"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "description": "User password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Request body for user login."
      },
...
      "PageResponse_CommitResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommitResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "PageResponse[CommitResponse]"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "minLength": 1,
            "title": "Refresh Token",
            "description": "A valid refresh token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "description": "Request body for token refresh."
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email",
            "description": "User email address",
            "examples": [
              "user@example.com"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "description": "Password (min 8 chars, must include uppercase, lowercase, digit)"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "description": "Display name",
            "examples": [
              "Alice Zhang"
            ]
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id",
            "description": "Team to join"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "name",
          "team_id"
        ],
        "title": "RegisterRequest",
        "description": "Request body for user registration."
      },
      "RegisterResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "User UUID"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User email"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Display name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "RegisterResponse",
        "description": "Response after successful registration."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "Short-lived JWT access token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token",
            "description": "Long-lived JWT refresh token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Token type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expires In",
            "description": "Access token expiry in seconds"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "description": "Response after successful login containing JWT tokens."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkloadResponse": {
        "properties": {
          "member_id": {
            "type": "string",
            "title": "Member Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "total_commits": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Commits"
          },
          "total_lines_added": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Lines Added"
          },
          "total_lines_deleted": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Lines Deleted"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          }
        },
        "type": "object",
        "required": [
          "member_id",
          "user_id",
          "total_commits",
          "total_lines_added",
          "total_lines_deleted"
        ],
        "title": "WorkloadResponse",
        "description": "Response schema for member workload statistics."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}